smart-learning-assistant/requirements.txt
langchain==0.2.16
langchain-google-genai
langchain-community
langchain-chroma
langserve[all]
httpx>=0.23.0,<0.28.0     # langserve 0.2.x requires VerifyTypes removed in httpx 0.28
fastapi
uvicorn[standard]
gradio
chromadb
pymupdf
pdfplumber
python-docx
sentence-transformers
opencv-python-headless
scikit-image
numpy
scipy
pandas
matplotlib
Pillow
python-dotenv
ragas
rouge-score
pytest
Nougat
